ms_management_client/src/requirements.txt
SQLAlchemy==2.0.41
python-dotenv==1.1.1
psycopg2-binary==2.9.10